Renamed image to shape all around.
