+++ src/aturtle/shapes/bitmap.py
