+++ src/aturtle/sprite/bitmap.py
@@ -10,14 +10,14 @@ from . import base
 
 class BitmapSprite(base.BaseSprite):
 
-    def __init__(self, canvas, image, *, x=0, y=0):
+    def __init__(self, canvas, shape, *, x=0, y=0):
 
-        super().__init__(canvas, image, x=x, y=y)
+        super().__init__(canvas, shape, x=x, y=y)
 
         self._id = self._canvas.create_image(
-            x - image.cx,
-            y - image.cy,
-            image=image[self._theta],
+            x - shape.cx,
+            y - shape.cy,
+            image=shape[self._theta],
             anchor='nw',
         )
 
@@ -27,16 +27,16 @@ class BitmapSprite(base.BaseSprite):
         # Rotate anchor point if needed.
         super().rotate(theta, around=around, update=False)
 
-        # Anchor point rotated, move the image.
+        # Anchor point rotated, move the shape.
         if around:
             self._canvas.moveto(
                 self._id,
-                self._x_anchor - self._image.cx,
-                self._y_anchor - self._image.cy,
+                self._x_anchor - self._shape.cx,
+                self._y_anchor - self._shape.cy,
             )
 
-        # Use the pre-rendered image for the new orientation.
-        self._canvas.itemconfig(self._id, image=self._image[self._theta])
+        # Use the pre-rendered shape for the new orientation.
+        self._canvas.itemconfig(self._id, image=self._shape[self._theta])
 
         if update:
             self.update()
